shoppingmall_server/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
environ==1.0
idna==3.4
kombu==5.3.5
oauthlib==3.2.2
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
